feat(#179): tsconfig paths 설정, cycle 에러 수정
