+++ src/assets/types/MyTravelItem.ts
@@ -1,0 +1,13 @@
+export interface TMyTravelItem {
+  id: number;
+  name: string;
+  departureDate: string;
+  arrivalDate: string;
+  location: string[];
+  days: {
+    day: number;
+    date: string;
+    dayOfWeek: string;
+  }[];
+  thumbnailURL: string;
+}
